+++ src/tercom/dao/TercomProfileDAO.php
@@ -4,31 +4,44 @@ namespace tercom\dao;
 
 use dProject\MySQL\Result;
 use dProject\Primitive\StringUtil;
-use tercom\Functions;
 use tercom\dao\exceptions\DAOException;
 use tercom\entities\TercomProfile;
 use tercom\entities\lists\TercomProfiles;
+use tercom\Functions;
 
 /**
+ * DAO para Perfil TERCOM
+ *
+ * Classe responsável pela comunicação completa do sistema para com o banco de dados.
+ * Sua responsabilidade é gerenciar os dados referentes aos perfis TERCOM, incluindo todas operações.
+ * Estas operações consiste em: adicionar, atualizar, selecionar e excluir <b>se não referenciados</b>.
+ *
+ * Perfis TERCOM possui apenas um nome que deve ser único e um nível de assinatura.
+ *
+ * @see GenericDAO
  * @see TercomProfile
+ * @see TercomProfiles
+ *
  * @author Andrew
  */
 class TercomProfileDAO extends GenericDAO
 {
 	/**
-	 * @var array
+	 * @var array nome das colunas da tabela de perfis TERCOM.
 	 */
 	public const ALL_COLUMNS = ['id', 'name', 'assignmentLevel'];
 
 	/**
-	 *
-	 * @param TercomProfile $tercomProfile
-	 * @param bool $validateID
-	 * @throws DAOException
+	 * Procedimento interno para validação dos dados de um perfil TERCOM ao inserir e/ou atualizar.
+	 * Perfis TERCOM não podem ter nome e nível de assinatura não informados.
+	 * @param TercomProfile $tercomProfile objeto do tipo perfil TERCOM à ser validado.
+	 * @param bool $validateId true para validar o código de identificação único ou false caso contrário.
+	 * @throws DAOException caso algum dos dados do perfil TERCOM não estejam de acordo.
 	 */
-	private function validate(TercomProfile $tercomProfile, bool $validateID)
+	private function validate(TercomProfile $tercomProfile, bool $validateId)
 	{
-		if ($validateID) {
+		// PRIMARY KEY
+		if ($validateId) {
 			if ($tercomProfile->getId() === 0)
 				throw new DAOException('perfil não identificado');
 		} else {
@@ -36,14 +49,15 @@ class TercomProfileDAO extends GenericDAO
 				throw new DAOException('perfil já identificado');
 		}
 
+		// NOT NULL
 		if (StringUtil::isEmpty($tercomProfile->getName())) throw new DAOException('nome não definido');
 		if ($tercomProfile->getAssignmentLevel() === 0) throw new DAOException('nível de assinatura não definido');
 	}
 
 	/**
-	 *
-	 * @param TercomProfile $tercomProfile
-	 * @return bool
+	 * Insere um novo perfil TERCOM no banco de dados e atualiza o mesmo com o identificador gerado.
+	 * @param TercomProfile $tercomProfile objeto do tipo perfil TERCOM à adicionar.
+	 * @return bool true se conseguir adicionar ou false caso contrário.
 	 */
 	public function insert(TercomProfile $tercomProfile): bool
 	{
@@ -63,9 +77,9 @@ class TercomProfileDAO extends GenericDAO
 	}
 
 	/**
-	 *
-	 * @param TercomProfile $tercomProfile
-	 * @return bool
+	 * Atualiza os dados de um perfil TERCOM já existente no banco de dados.
+	 * @param TercomProfile $tercomProfile objeto do tipo perfil TERCOM à atualizar.
+	 * @return bool true se for atualizado ou false caso contrário.
 	 */
 	public function update(TercomProfile $tercomProfile): bool
 	{
@@ -84,13 +98,17 @@ class TercomProfileDAO extends GenericDAO
 	}
 
 	/**
-	 *
-	 * @param TercomProfile $tercomProfile
-	 * @return bool
+	 * Exclui os dados de um perfil TERCOM já existente no banco de dados.
+	 * @param TercomProfile $tercomProfile objeto do tipo perfil TERCOM à excluir.
+	 * @return bool true se for excluído ou false caso contrário.
+	 * @throws DAOException um ou mais funcionários definidos no perfil.
 	 */
 	public function delete(TercomProfile $tercomProfile): bool
 	{
 		$this->validate($tercomProfile, true);
+
+		if ($this->existOnTercomEmployees($tercomProfile))
+			throw new DAOException('um ou mais funcionários definido no perfil');
 
 		$sql = "DELETE FROM tercom_profiles
 				WHERE id = ?";
@@ -102,8 +120,8 @@ class TercomProfileDAO extends GenericDAO
 	}
 
 	/**
-	 *
-	 * @return string
+	 * Procedimento interno para centralizar e agilizar a manutenção de queries.
+	 * @return string aquisição da string de consulta simples para SELECT.
 	 */
 	private function newBaseSelect(): string
 	{
@@ -112,9 +130,9 @@ class TercomProfileDAO extends GenericDAO
 	}
 
 	/**
-	 *
-	 * @param int $idTercomProfile
-	 * @return TercomProfile|NULL
+	 * Selecione os dados de um perfil TERCOM através do seu código de identificação único.
+	 * @param int $idTercomProfile código de identificação único do perfil TERCOM.
+	 * @return TercomProfile|NULL perfil TERCOM com os dados carregados ou NULL se não encontrado.
 	 */
 	public function select(int $idTercomProfile): ?TercomProfile
 	{
@@ -131,9 +149,9 @@ class TercomProfileDAO extends GenericDAO
 	}
 
 	/**
-	 *
-	 * @param int $assignmentLevel
-	 * @return TercomProfiles|NULL
+	 * Seleciona os dados de um perfil TERCOM através do seu nível de assinatura.
+	 * @param int $assignmentLevel nível de assinatura máximo à filtrar.
+	 * @return TercomProfiles aquisição da lista de perfil TERCOM conforme filtro.
 	 */
 	public function selectByAssignmentLevel(int $assignmentLevel): TercomProfiles
 	{
@@ -150,8 +168,8 @@ class TercomProfileDAO extends GenericDAO
 	}
 
 	/**
-	 *
-	 * @return TercomProfiles|NULL
+	 * Seleciona os dados de todos os perfil TERCOM registrados no banco de dados sem ordenação.
+	 * @return TercomProfiles aquisição da lista de perfil TERCOM atualmente registrados.
 	 */
 	public function selectAll(): TercomProfiles
 	{
@@ -164,9 +182,9 @@ class TercomProfileDAO extends GenericDAO
 	}
 
 	/**
-	 *
-	 * @param int $idTercomProfile
-	 * @return bool
+	 * Verifica se um determinado código de identificação de perfil TERCOM existe.
+	 * @param int $idTercomProfile código de identificação único do perfil TERCOM.
+	 * @return bool true se existir ou false caso contrário.
 	 */
 	public function exist(int $idTercomProfile): bool
 	{
@@ -177,18 +195,15 @@ class TercomProfileDAO extends GenericDAO
 		$query = $this->createQuery($sql);
 		$query->setInteger(1, $idTercomProfile);
 
-		$result = $query->execute();
-		$entry = $result->next();
-		$result->free();
-
-		return intval($entry['qty']) === 1;
+		return $this->parseQueryExist($query);
 	}
 
 	/**
-	 *
-	 * @param string $name
-	 * @param int $idTercomProfile
-	 * @return bool
+	 * Verifica se um determinado nome de perfil está disponível para perfil TERCOM.
+	 * @param string $name nome do perfil à verificar.
+	 * @param int $idTercomProfile código de identificação do perfil TERCOM à desconsiderar
+	 * ou zero caso seja um novo perfil TERCOM.
+	 * @return bool true se existir ou false caso contrário.
 	 */
 	public function existName(string $name, int $idTercomProfile): bool
 	{
@@ -200,14 +215,30 @@ class TercomProfileDAO extends GenericDAO
 		$query->setString(1, $name);
 		$query->setInteger(2, $idTercomProfile);
 
-		$result = $query->execute();
-		$entry = $result->next();
-		$result->free();
-
-		return intval($entry['qty']) === 1;
+		return $this->parseQueryExist($query);
 	}
 
 	/**
+	 * Verifica se um perfil TERCOM está sendo referenciado em algum funcionário TERCOM.
+	 * @param TercomProfile $tercomProfile perfil TERCOM à ser verificar.
+	 * @return bool true se existir ou false caso contrário.
+	 */
+	public function existOnTercomEmployees(TercomProfile $tercomProfile): bool
+	{
+		$sql = "SELECT COUNT(*) qty
+				FROM tercom_employees
+				WHERE idTercomProfile = ?";
+
+		$query = $this->createQuery($sql);
+		$query->setInteger(1, $tercomProfile->getId());
+
+		return $this->parseQueryExist($query);
+	}
+
+	/**
+	 * Procedimento interno para analisar o resultado de uma consulta e criar um objeto de perfil TERCOM.
+	 * @param Result $result referência do resultado da consulta obtido.
+	 * @return TercomProfile|NULL objeto do tipo perfil TERCOM com dados carregados ou NULL se não houver resultado.
 	 *
 	 * @param Result $result
 	 * @return TercomProfile|NULL
@@ -218,6 +249,9 @@ class TercomProfileDAO extends GenericDAO
 	}
 
 	/**
+	 * Procedimento interno para analisar o resultado de uma consulta e criar os objetos de perfil TERCOM.
+	 * @param Result $result referência do resultado da consulta obtido.
+	 * @return TercomProfiles aquisição da lista de perfis TERCOM a partir da consulta.
 	 *
 	 * @param Result $result
 	 * @return TercomProfiles
@@ -236,9 +270,9 @@ class TercomProfileDAO extends GenericDAO
 	}
 
 	/**
-	 *
-	 * @param array $entry
-	 * @return TercomProfile
+	 * Procedimento interno para criar um objeto do tipo perfil TERCOM e carregar os dados de um registro.
+	 * @param array $entry vetor contendo os dados do registro obtido de uma consulta.
+	 * @return TercomProfile aquisição de um objeto do tipo perfil TERCOM com dados carregados.
 	 */
 	private function newTercomProfile(array $entry): TercomProfile
 	{
